feat: Add terminal extension with capabilities for initialization, writing, resizing, closing, and saving.
