+++ src-tauri/src/commands/mod.rs
@@ -3,6 +3,7 @@ pub mod checkout;
 pub mod editor;
 pub mod event;
 pub mod file;
+pub mod terminal;
 
 // Re-export all commands for easy registration
 pub use block::{check_permission, execute_command, get_all_blocks, get_block, rename_block};
@@ -12,3 +13,4 @@ pub use file::{
     close_file, create_file, get_all_events, get_file_info, list_open_files, open_file,
     rename_file, save_file, FileMetadata,
 };
+pub use terminal::{async_init_terminal, close_terminal_session, resize_pty, write_to_pty};
