+++ src-tauri/src/extensions/terminal/terminal_resize.rs
@@ -1,70 +1,42 @@
-//! Terminal resize command
+//! terminal.resize Capability
 //!
-//! Resizes an active PTY session window.
+//! Authorization capability for terminal resize operations.
+//! The actual PTY resize is handled by the Tauri command in `commands/terminal.rs`.
 
-use portable_pty::PtySize;
-use specta::specta;
-use tauri::State;
+use crate::capabilities::core::CapResult;
+use crate::models::{Block, Command, Event};
+use capability_macros::capability;
 
-use super::state::TerminalState;
-use super::TerminalResizePayload;
-use crate::state::AppState;
-
-/// Resize the PTY.
+/// Handler for terminal.resize capability.
 ///
-/// This command updates the PTY window dimensions when the frontend
-/// terminal viewport changes size.
+/// This capability performs authorization check for resizing a terminal.
+/// It acts as a permission gate - if this handler succeeds, the caller is
+/// authorized to resize the PTY session.
+///
+/// # Architecture Note
+/// In Elfiee's split architecture:
+/// 1. **Capability Handler** (this file): Performs authorization check
+/// 2. **Tauri Command** (`commands/terminal.rs`): Performs actual PTY resize
 ///
 /// # Arguments
-/// * `state` - Terminal state containing active sessions
-/// * `app_state` - Application state for permission checking
-/// * `payload` - Resize parameters (cols, rows, block_id, file_id, editor_id)
+/// * `cmd` - The command containing resize parameters
+/// * `block` - The terminal block (must exist and be of type "terminal")
 ///
 /// # Returns
-/// * `Ok(())` - PTY resized successfully
-/// * `Err(String)` - Error if session not found or resize fails
-#[tauri::command]
-#[specta]
-pub async fn resize_pty(
-    state: State<'_, TerminalState>,
-    app_state: State<'_, AppState>,
-    payload: TerminalResizePayload,
-) -> Result<(), String> {
-    // Verify permissions using capability system via EngineHandle
-    let engine = app_state
-        .engine_manager
-        .get_engine(&payload.file_id)
-        .ok_or_else(|| format!("File '{}' is not open", payload.file_id))?;
+/// * `Ok(Vec<Event>)` - Empty vector (authorization passed, no events generated)
+/// * `Err(String)` - Error if block is missing or invalid
+#[capability(id = "terminal.resize", target = "terminal")]
+fn handle_terminal_resize(_cmd: &Command, block: Option<&Block>) -> CapResult<Vec<Event>> {
+    let block = block.ok_or("Block required for terminal.resize")?;
 
-    let authorized = engine
-        .check_grant(
-            payload.editor_id.clone(),
-            "terminal.resize".to_string(),
-            payload.block_id.clone(),
-        )
-        .await;
-
-    if !authorized {
+    if block.block_type != "terminal" {
         return Err(format!(
-            "Authorization failed: {} does not have permission for terminal.resize on block {}",
-            payload.editor_id, payload.block_id
+            "Expected block_type 'terminal', got '{}'",
+            block.block_type
         ));
     }
 
-    let mut sessions = state.sessions.lock().unwrap();
-    if let Some(session) = sessions.get_mut(&payload.block_id) {
-        session
-            .master
-            .resize(PtySize {
-                rows: payload.rows,
-                cols: payload.cols,
-                pixel_width: 0,
-                pixel_height: 0,
-            })
-            .map_err(|e| format!("Failed to resize PTY: {}", e))?;
-    } else {
-        return Err("Terminal session not found".to_string());
-    }
-
-    Ok(())
+    // Authorization is handled by the engine before calling this handler.
+    // Return empty events - no state change for resize permission check.
+    Ok(vec![])
 }
